Renamed resources.py to base.py to match other providers
