+++ ScoutSuite/providers/gcp/resources/base.py
