+++ ScoutSuite/providers/gcp/resources/cloudsql/database_instances.py
@@ -1,6 +1,6 @@
 import asyncio
 from ScoutSuite.providers.gcp.facade.gcp import GCPFacade
-from ScoutSuite.providers.gcp.resources.resources import GCPCompositeResources
+from ScoutSuite.providers.gcp.resources.base import GCPCompositeResources
 from ScoutSuite.providers.gcp.resources.cloudsql.backups import Backups
 from ScoutSuite.providers.gcp.resources.cloudsql.users import Users
 from ScoutSuite.providers.utils import get_non_provider_id
